+++ ops/src/lib.rs
@@ -8,10 +8,10 @@ pub mod force_push;
 pub mod git;
 pub mod git_clone_manager;
 pub mod messenger;
+pub mod migrate_slack;
 pub mod pr_merge;
 pub mod repo_version;
 pub mod slack;
-pub mod users_migrate;
 pub mod util;
 pub mod webhook_db;
 mod webhook_db_migrations;
